+++ app/widgets/NewsNav/NewsNav.php
@@ -6,8 +6,14 @@ include_once WIDGETS_PATH . 'Post/Post.php';
 
 class NewsNav extends Base
 {
-    public function display()
+    public function load()
     {
+        $this->addjs('newsnav.js');
+    }
+
+    public function ajaxHttpGet($page, $server)
+    {
+        $view = $this->tpl();
         $blogs = \App\Post::where('open', true)
                           ->orderBy('posts.published', 'desc')
                           ->restrictToMicroblog()
@@ -18,7 +24,7 @@ class NewsNav extends Base
                           ->get()
                           ->shuffle();
 
-        $this->view->assign('blogs', $blogs);
+        $view->assign('blogs', $blogs);
 
         $posts = \App\Post::where('open', true)
                           ->orderBy('posts.published', 'desc')
@@ -28,11 +34,14 @@ class NewsNav extends Base
                           ->recents()
                           ->take(6);
 
-        if ($this->get('s') && $this->get('s') != 'subscriptions') {
-            $posts->where('posts.server', $this->get('s'));
+        if (isset($server) && $server != 'subscriptions') {
+            $posts->where('posts.server', $server);
         }
 
-        $this->view->assign('posts', $posts->get()->shuffle());
+        $view->assign('posts', $posts->get()->shuffle());
+        $view->assign('page', $page);
+
+        $this->rpc('MovimTpl.fill', '#newsnav', $view->draw('_newsnav'));
     }
 
     public function prepareTicket(\App\Post $post)