+++ app/widgets/NewsNav/newsnav.js
@@ -1,0 +1,4 @@
+MovimWebsocket.initiate(() => NewsNav_ajaxHttpGet(
+    MovimUtils.urlParts().page,
+    MovimUtils.urlParts().params[0]
+));