+++ src/Movim/Bootstrap.php
@@ -187,6 +187,10 @@ class Bootstrap
 
             $dispatcher = new Dispatcher(new Container);
             $dispatcher->listen('Illuminate\Database\Events\QueryExecuted', function ($query) use (&$sqlQueryExecuted) {
+                if (strstr($query->sql, 'presences')) {
+                    \Utils::error($query->sql);
+                    \Utils::error(serialize($query->bindings));
+                }
                 $sqlQueryExecuted = time();
             });