+++ CHANGELOG.md
@@ -18,6 +18,7 @@ v0.18.1 (trunk)
  * Add ~1500 new emojis to the Javascript selector
  * Implement XEP-0201: Best Practices for Message Threads and add Reply feature
  * Add a cache for translations, refreshed when the daemon is launched or by a explicit command
+ * Make NewsNav asynchrone
 
 v0.18
 ---------------------------